hayaku_voice_input/requirements.txt
PyQt5>=5.15.0
pyaudio
webrtcvad-wheels
numpy
pydub
openai
requests
pyyaml
